# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Task Manager API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TaskPriority = (typeof TaskPriority)[keyof typeof TaskPriority];

export const TaskPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  urgent: "urgent",
} as const;

export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus];

export const TaskStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Task {
  id: number;
  title: string;
  description?: string | null;
  categoryId?: number | null;
  categoryName?: string | null;
  categoryColor?: string | null;
  priority: TaskPriority;
  status: TaskStatus;
  dueDate?: string | null;
  dueTime?: string | null;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateTaskRequestPriority =
  (typeof CreateTaskRequestPriority)[keyof typeof CreateTaskRequestPriority];

export const CreateTaskRequestPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  urgent: "urgent",
} as const;

export interface CreateTaskRequest {
  title: string;
  description?: string | null;
  categoryId?: number | null;
  priority: CreateTaskRequestPriority;
  dueDate?: string | null;
  dueTime?: string | null;
  notes?: string | null;
}

export type UpdateTaskRequestPriority =
  (typeof UpdateTaskRequestPriority)[keyof typeof UpdateTaskRequestPriority];

export const UpdateTaskRequestPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  urgent: "urgent",
} as const;

export type UpdateTaskRequestStatus =
  (typeof UpdateTaskRequestStatus)[keyof typeof UpdateTaskRequestStatus];

export const UpdateTaskRequestStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface UpdateTaskRequest {
  title?: string;
  description?: string | null;
  categoryId?: number | null;
  priority?: UpdateTaskRequestPriority;
  status?: UpdateTaskRequestStatus;
  dueDate?: string | null;
  dueTime?: string | null;
  notes?: string | null;
}

export interface RescheduleTaskRequest {
  dueDate: string | null;
  dueTime?: string | null;
}

export type TasksSummaryByPriority = {
  low: number;
  medium: number;
  high: number;
  urgent: number;
};

export interface TasksSummary {
  total: number;
  pending: number;
  inProgress: number;
  completed: number;
  cancelled: number;
  overdue: number;
  dueToday: number;
  dueSoon: number;
  byPriority: TasksSummaryByPriority;
}

export interface Category {
  id: number;
  name: string;
  color: string;
  icon?: string | null;
  taskCount: number;
  createdAt: string;
}

export interface CreateCategoryRequest {
  name: string;
  color: string;
  icon?: string | null;
}

export interface MessageResponse {
  message: string;
}

export interface ErrorResponse {
  error: string;
}

export type GetTasksParams = {
  category?: string;
  status?: GetTasksStatus;
  priority?: GetTasksPriority;
  dueDate?: string;
};

export type GetTasksStatus =
  (typeof GetTasksStatus)[keyof typeof GetTasksStatus];

export const GetTasksStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type GetTasksPriority =
  (typeof GetTasksPriority)[keyof typeof GetTasksPriority];

export const GetTasksPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  urgent: "urgent",
} as const;