# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Task Manager API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: tasks
    description: Task management operations
  - name: categories
    description: Category management operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /tasks:
    get:
      operationId: getTasks
      tags: [tasks]
      summary: List all tasks
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, in_progress, completed, cancelled]
        - name: priority
          in: query
          required: false
          schema:
            type: string
            enum: [low, medium, high, urgent]
        - name: dueDate
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /tasks/{id}/complete:
    post:
      operationId: completeTask
      tags: [tasks]
      summary: Mark a task as complete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task marked complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /tasks/{id}/reschedule:
    post:
      operationId: rescheduleTask
      tags: [tasks]
      summary: Reschedule a task to a new due date
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RescheduleTaskRequest"
      responses:
        "200":
          description: Task rescheduled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /tasks/summary:
    get:
      operationId: getTasksSummary
      tags: [tasks]
      summary: Get task summary statistics
      responses:
        "200":
          description: Summary statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TasksSummary"

  /tasks/upcoming:
    get:
      operationId: getUpcomingTasks
      tags: [tasks]
      summary: Get tasks due in the next 7 days
      responses:
        "200":
          description: Upcoming tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"

  /tasks/overdue:
    get:
      operationId: getOverdueTasks
      tags: [tasks]
      summary: Get overdue tasks
      responses:
        "200":
          description: Overdue tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"

  /categories:
    get:
      operationId: getCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a new category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

  /categories/{id}:
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Category not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        categoryId:
          type: integer
          nullable: true
        categoryName:
          type: string
          nullable: true
        categoryColor:
          type: string
          nullable: true
        priority:
          type: string
          enum: [low, medium, high, urgent]
        status:
          type: string
          enum: [pending, in_progress, completed, cancelled]
        dueDate:
          type: string
          nullable: true
        dueTime:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - priority
        - status
        - createdAt
        - updatedAt

    CreateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        categoryId:
          type: integer
          nullable: true
        priority:
          type: string
          enum: [low, medium, high, urgent]
        dueDate:
          type: string
          nullable: true
        dueTime:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - title
        - priority

    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        categoryId:
          type: integer
          nullable: true
        priority:
          type: string
          enum: [low, medium, high, urgent]
        status:
          type: string
          enum: [pending, in_progress, completed, cancelled]
        dueDate:
          type: string
          nullable: true
        dueTime:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true

    RescheduleTaskRequest:
      type: object
      properties:
        dueDate:
          type: string
          nullable: true
        dueTime:
          type: string
          nullable: true
      required:
        - dueDate

    TasksSummary:
      type: object
      properties:
        total:
          type: integer
        pending:
          type: integer
        inProgress:
          type: integer
        completed:
          type: integer
        cancelled:
          type: integer
        overdue:
          type: integer
        dueToday:
          type: integer
        dueSoon:
          type: integer
        byPriority:
          type: object
          properties:
            low:
              type: integer
            medium:
              type: integer
            high:
              type: integer
            urgent:
              type: integer
          required:
            - low
            - medium
            - high
            - urgent
      required:
        - total
        - pending
        - inProgress
        - completed
        - cancelled
        - overdue
        - dueToday
        - dueSoon
        - byPriority

    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
        icon:
          type: string
          nullable: true
        taskCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - color
        - taskCount
        - createdAt

    CreateCategoryRequest:
      type: object
      properties:
        name:
          type: string
        color:
          type: string
        icon:
          type: string
          nullable: true
      required:
        - name
        - color

    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error