# lib/api-zod/src/generated/types/category.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Task Manager API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Category {
  id: number;
  name: string;
  color: string;
  icon?: string | null;
  taskCount: number;
  createdAt: string;
}