# lib/api-zod/src/generated/types/createCategoryRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Task Manager API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCategoryRequest {
  name: string;
  color: string;
  icon?: string | null;
}