# lib/api-zod/src/generated/types/rescheduleTaskRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Task Manager API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RescheduleTaskRequest {
  dueDate: string | null;
  dueTime?: string | null;
}