# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Task Manager API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TaskPriority } from "./taskPriority";
import type { TaskStatus } from "./taskStatus";

export interface Task {
  id: number;
  title: string;
  description?: string | null;
  categoryId?: number | null;
  categoryName?: string | null;
  categoryColor?: string | null;
  priority: TaskPriority;
  status: TaskStatus;
  dueDate?: string | null;
  dueTime?: string | null;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}