# lib/api-zod/src/generated/types/tasksSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Task Manager API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TasksSummaryByPriority } from "./tasksSummaryByPriority";

export interface TasksSummary {
  total: number;
  pending: number;
  inProgress: number;
  completed: number;
  cancelled: number;
  overdue: number;
  dueToday: number;
  dueSoon: number;
  byPriority: TasksSummaryByPriority;
}