# lib/api-zod/src/generated/types/updateTaskRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Task Manager API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateTaskRequestPriority } from "./updateTaskRequestPriority";
import type { UpdateTaskRequestStatus } from "./updateTaskRequestStatus";

export interface UpdateTaskRequest {
  title?: string;
  description?: string | null;
  categoryId?: number | null;
  priority?: UpdateTaskRequestPriority;
  status?: UpdateTaskRequestStatus;
  dueDate?: string | null;
  dueTime?: string | null;
  notes?: string | null;
}